Switching to FSD - END
Переход на архитектуру FSD
