--- src/components/accountPage/devicesInfo/DevicesInfo.module.scss
@@ -1,9 +1,0 @@
-.devices {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  &__title {
-    font-size: 24px;
-    font-weight: 700;
-  }
-}
