--- src/components/accountPage/devicesInfo/productInfo/CardDeviceItem.module.scss
@@ -1,56 +1,0 @@
-.card {
-  border: 1px solid rgba(255, 255, 255, 0.16);
-  border-radius: 8px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  color: #fff;
-
-  width: 300px;
-
-  &__content {
-    display: flex;
-    gap: 16px;
-    padding: 16px;
-  }
-
-  &__head {
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-    justify-content: space-between;
-    width: 100%;
-  }
-  &__title, &__price {
-    font-weight: bold;
-  }
-  
-  &__subtitle {
-    opacity: 0.5;
-  }
-
-  &__img {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    outline: 1px solid rgba($color: white, $alpha: 0.16);
-    border-radius: 4px;
-
-    overflow: hidden;
-    width: 100%;
-    height: 300px;
-    .img {
-      width: 100%;
-      height: 100%;
-      object-fit: cover;
-    }
-  }
-  &__button {
-    display: flex;
-    align-self: flex-end;
-    border-radius: 4px;
-    width: 42px;
-    height: 42px;
-  }
-}
