--- src/components/accountPage/devicesInfo/productRegistrationInfo/ProductRegistrationInfo.module.scss
@@ -1,16 +1,0 @@
-.productRegistration {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  &__head {
-    display: flex;
-    flex-direction: column;
-  }
-  &__title {
-    font-weight: 700;
-  }
-  &__subtitle {
-    opacity: 0.5;
-  }
-}
