+++ src/entities​/Categories/model/index.ts
@@ -1,0 +1,1 @@
+export { }
