+++ src/entities​/Categories/ui/CategoriesCard/CategoriesCard.module.scss
@@ -1,0 +1,46 @@
+.card {
+    outline: 1px solid rgba(255, 255, 255, 0.16);
+    border-radius: 8px;
+    // padding: 8px;
+    display: flex;
+    justify-content: space-between;
+    gap: 16px;
+    cursor: pointer;
+    color: #fff;
+    min-height: 200px;
+    max-height: 200px;
+    transition: outline 0.1s ease-out;
+    overflow: hidden;
+  
+    &__content {
+      padding: 16px;
+      display: flex;
+      flex-direction: column;
+      gap: 16px;
+    }
+    &__title {
+      font-size: 16px;
+      font-weight: 700;
+    }
+    &__subtitle {
+    }
+    &__description {
+      opacity: 0.5;
+    }
+    &__wrapper {
+      border-radius: 8px;
+      min-width: 60%;
+      background: rgba($color: #fff, $alpha: 0.3);
+      overflow: hidden;
+    }
+    &__img {
+      object-fit:cover;
+      width: 100%;
+      height: 100%;
+    }
+    &:hover {
+      outline: 2px solid #f77f00;
+      outline-offset: 4px;
+    }
+  }
+
