+++ src/entities​/Categories/ui/CategoriesCard/CategoriesCardInterface.ts
@@ -1,0 +1,8 @@
+export interface Categories {
+    title: string;
+    subtitle: string;
+    description: string;
+    imageUrl: string;
+    altText: string;
+  }
+
