+++ src/entities​/Categories/ui/index.ts
@@ -1,0 +1,3 @@
+import CategoriesCard from './CategoriesCard/CategoriesCard.tsx'
+
+export { CategoriesCard }
