+++ src/entities​/Product/index.ts
@@ -1,0 +1,2 @@
+export * from './model'
+export * from './ui'
