+++ src/entities​/Product/ui/ProductCard/ProductCardInterface.ts
@@ -1,0 +1,8 @@
+export interface Product {
+    title: string;
+    subtitle: string;
+    price: string;
+    imageUrl: string;
+    altText: string;
+    onClick: () => void;
+  }
