+++ src/entities​/Product/ui/ProductOrderCard/ProductOrderCardInterface.ts
@@ -1,0 +1,20 @@
+import { Product } from "../ProductCard/ProductCardInterface";
+import { Categories } from "../../../Categories/ui/CategoriesCard/CategoriesCardInterface";
+import { Subscription } from "../../../Subscription/ui/SubscriptionCard/SubscriptionCardInterface";
+
+export interface CardUpProps {
+    visible: boolean;
+    selectedProduct: Product; 
+    selectedCategory: Categories; 
+    selectedSubscription: Subscription;
+    totalPrice: number;
+    getImage: (fileName: string) => string;
+  }
+  
+  export interface CardDownProps {
+    deliveryMethod: string;
+    onDeliveryMethodChange: (method: string) => void;
+    selectedProduct: Product;
+    selectedSubscription: Subscription;
+  }
+
