+++ src/entities​/Product/ui/index.ts
@@ -1,0 +1,4 @@
+import ProductCard from './ProductCard/ProductCard.tsx'
+import ProductOrderCard from './ProductOrderCard/ProductOrderCard.tsx'
+
+export { ProductCard, ProductOrderCard }
