+++ src/entities​/index.ts
@@ -3,4 +3,6 @@ export { StepsCardItemMain } from "./StepsCardItemMain";
 export { LineItemMain } from "./LineItemMain";
 
 export { SubscriptionCard } from "./Subscription";
+export { CategoriesCard } from "./Categories";
+export { ProductCard, ProductOrderCard } from "./Product";
 export { PaymentsCard, PaymentsPagination, PaymentsTable } from "./Payments";
