+++ src/shared/lib/utils/getCategoriesForm/getCategoryForm.tsx
