+++ src/widgets/Categories/ui/index.ts
@@ -1,0 +1,3 @@
+import Categories from './Categories.tsx'
+
+export { Categories }
