+++ src/widgets/Devices/ui/Devices.module.scss
@@ -1,0 +1,10 @@
+.devices {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+    &__title {
+      font-size: 24px;
+      font-weight: 700;
+    }
+  }
+
