+++ src/widgets/Devices/ui/index.ts
@@ -1,0 +1,3 @@
+import Devices from './Devices.tsx'
+
+export { Devices }
