+++ src/widgets/Products/ui/Product/Product.module.scss
