+++ src/widgets/Products/ui/ProductOrder/ProductOrder.module.scss
@@ -1,0 +1,17 @@
+.productRegistration {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+  
+    &__head {
+      display: flex;
+      flex-direction: column;
+    }
+    &__title {
+      font-weight: 700;
+    }
+    &__subtitle {
+      opacity: 0.5;
+    }
+  }
+
