+++ src/widgets/Products/ui/index.ts
@@ -1,0 +1,5 @@
+import Product from "./Product/Products";
+import ProductDetailed from "./ProductDetailed/ProductDetailed";
+import ProductOrder from "./ProductOrder/ProductOrder";
+
+export { Product, ProductDetailed, ProductOrder };
